third_party/blink/web_tests/editing/selection/DOMSelection-crossing-document-expected.txt

Test to check if setBaseAndExtent guard node with null owner document (Bug 31680)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS foreignSel.anchorNode is null
PASS mainSel.anchorNode is null
PASS foreignSel.anchorNode is null
PASS mainSel.anchorNode is null
PASS foreignSel.anchorNode is null
PASS mainSel.anchorNode is null
PASS foreignSel.anchorNode is null
PASS mainSel.anchorNode is null
PASS foreignSel.anchorNode is null
PASS mainSel.anchorNode is null
PASS foreignSel.anchorNode is null
PASS mainSel.anchorNode is null
PASS successfullyParsed is true

TEST COMPLETE

Main Text
